# frontend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: MTN Bulk Data Wholesale Service API
  version: 0.1.0
  description: Backend contract required by the frontend-only MTN Bulk Data Wholesale Service console.
  x-api-requirements:
    - Every response must use the standard success/error envelope documented in components.schemas.ApiEnvelope and components.schemas.ErrorEnvelope.
    - Authenticated operations require the hardened HTTP-only JWT session cookie unless the operation explicitly sets security to an empty list.
    - Paginated table endpoints must return page, limit, total, totalPages, hasNextPage, and hasPreviousPage in meta.
    - Phone and MSISDN input must be normalized to E.164 Uganda numbers matching +256(77|78|79|76|39)XXXXXXX.
    - List endpoints with search parameters must support backend search because the frontend uses asynchronous API search when local table search has no matches.
    - Payment initiation must return a pending transaction and payment session that can be tracked globally until confirmation, expiry, or failure.
    - Customer account activation must follow the solution-document flow: activation link, email OTP verification, password creation, activation notification, then immediate portal access.
    - Bundle package validity is fixed at 30 days; configurable bundle fields are service code, name, volume up to 4 TB, price, status, and visibility.
...
servers:
  - url: https://backend.example.com/api
    description: Live backend base URL
security:
  - cookieAuth: []
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customer accounts
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Search"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, deactivated, pending]
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: Customers fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
              examples:
                paginatedCustomers:
                  $ref: "#/components/examples/CustomerListResponseExample"
    post:
      tags: [Customers]
      summary: Register a customer
      description: Validates the primary MSISDN APN, adds subscriber to Provisioning, and triggers account activation notifications.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegistrationRequest"
            examples:
              registerCustomer:
                $ref: "#/components/examples/CustomerRegistrationRequestExample"
      responses:
        "200":
          description: Customer registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerRegistrationResponse"
              examples:
                registeredCustomer:
                  $ref: "#/components/examples/CustomerRegistrationResponseExample"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /customers/{customerId}:
    parameters:
      - $ref: "#/components/parameters/CustomerId"
    get:
      tags: [Customers]
      summary: Get customer account
      responses:
        "200":
          description: Customer fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
              examples:
                customer:
                  $ref: "#/components/examples/CustomerResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [Customers]
      summary: Update customer account details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdateRequest"
            examples:
              updateCustomer:
                $ref: "#/components/examples/CustomerUpdateRequestExample"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
              examples:
                updatedCustomer:
                  $ref: "#/components/examples/CustomerResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /customers/{customerId}/status:
    parameters:
      - $ref: "#/components/parameters/CustomerId"
    post:
      tags: [Customers]
      summary: Deactivate or reactivate customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerStatusChangeRequest"
            examples:
              deactivateCustomer:
                $ref: "#/components/examples/CustomerStatusChangeRequestExample"
      responses:
        "200":
          description: Customer status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
              examples:
                statusUpdated:
                  $ref: "#/components/examples/CustomerDeactivatedResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /customers/{customerId}/primary-msisdns:
    parameters:
      - $ref: "#/components/parameters/CustomerId"
    post:
      tags: [Customers]
      summary: Add primary MSISDN
      description: Runs APN validation through the configured VasApps/UDM provider before adding the primary MSISDN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrimaryMsisdnRequest"
            examples:
              addPrimaryMsisdn:
                $ref: "#/components/examples/PrimaryMsisdnRequestExample"
      responses:
        "200":
          description: Primary MSISDN added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrimaryMsisdnResponse"
              examples:
                primaryMsisdnAdded:
                  $ref: "#/components/examples/PrimaryMsisdnResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /customers/{customerId}/primary-msisdns/{primaryMsisdn}/balance:
    parameters:
      - $ref: "#/components/parameters/CustomerId"
      - $ref: "#/components/parameters/PrimaryMsisdn"
    get:
      tags: [Customers]
      summary: Check bundle balance
      description: Retrieves bundle volume and expiry from Provisioning.
      responses:
        "200":
          description: Balance fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceResponse"
              examples:
                balance:
                  $ref: "#/components/examples/BalanceResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /customers/{customerId}/primary-msisdns/{primaryMsisdn}/secondary-numbers:
    parameters:
      - $ref: "#/components/parameters/CustomerId"
      - $ref: "#/components/parameters/PrimaryMsisdn"
    get:
      tags: [Secondary Numbers]
      summary: List secondary numbers
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Search"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, removed]
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: Secondary numbers fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecondaryNumberListResponse"
              examples:
                paginatedSecondaryNumbers:
                  $ref: "#/components/examples/SecondaryNumberListResponseExample"
    post:
      tags: [Secondary Numbers]
      summary: Add secondary number
      description: Validates the secondary MSISDN through the configured VasApps/UDM APN provider before adding the number. The request is rejected when the provider returns multiple APNs or a single APN that does not match the registered customer APN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecondaryNumberRequest"
            examples:
              addSecondaryNumber:
                $ref: "#/components/examples/SecondaryNumberRequestExample"
      responses:
        "200":
          description: Secondary number added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecondaryNumberMutationResponse"
              examples:
                secondaryNumberAdded:
                  $ref: "#/components/examples/SecondaryNumberMutationResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /customers/{customerId}/primary-msisdns/{primaryMsisdn}/secondary-numbers/bulk:
    parameters:
      - $ref: "#/components/parameters/CustomerId"
      - $ref: "#/components/parameters/PrimaryMsisdn"
    post:
      tags: [Secondary Numbers]
      summary: Bulk upload secondary numbers
      description: Validates each uploaded MSISDN through the configured VasApps/UDM APN provider and reports rejected numbers with the exact APN rule that failed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecondaryBulkRequest"
            examples:
              bulkUploadSecondaryNumbers:
                $ref: "#/components/examples/SecondaryBulkRequestExample"
      responses:
        "200":
          description: Bulk upload processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecondaryBulkResponse"
              examples:
                bulkUploadProcessed:
                  $ref: "#/components/examples/SecondaryBulkResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /customers/{customerId}/primary-msisdns/{primaryMsisdn}/secondary-numbers/{secondaryMsisdn}:
    parameters:
      - $ref: "#/components/parameters/CustomerId"
      - $ref: "#/components/parameters/PrimaryMsisdn"
      - name: secondaryMsisdn
        in: path
        required: true
        description: URL-encode the leading plus sign when sending this path parameter, for example %2B256772991001.
        schema:
          $ref: "#/components/schemas/UgandaPhoneNumber"
    delete:
      tags: [Secondary Numbers]
      summary: Remove secondary number
      description: Sends deleteGroupMember to Provisioning and removes the local secondary-number flag.
      responses:
        "200":
          description: Secondary number removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecondaryNumberResponse"
              examples:
                secondaryNumberRemoved:
                  $ref: "#/components/examples/SecondaryNumberRemovedResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /customers/{customerId}/primary-msisdns/{primaryMsisdn}/secondary-numbers/{secondaryMsisdn}/usage:
    parameters:
      - $ref: "#/components/parameters/CustomerId"
      - $ref: "#/components/parameters/PrimaryMsisdn"
      - name: secondaryMsisdn
        in: path
        required: true
        description: URL-encode the leading plus sign when sending this path parameter, for example %2B256772991001.
        schema:
          $ref: "#/components/schemas/UgandaPhoneNumber"
    get:
      tags: [Secondary Numbers]
      summary: Check secondary number usage
      description: Returns usage and remaining allocation for a secondary MSISDN under a customer primary MSISDN.
      responses:
        "200":
          description: Secondary MSISDN usage fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecondaryUsageResponse"
              examples:
                secondaryUsage:
                  $ref: "#/components/examples/SecondaryUsageResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /bundles:
    get:
      tags: [Bundles]
      summary: List configured wholesale bundles
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, paused, disabled]
        - name: visible
          in: query
          required: false
          schema:
            type: boolean
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Bundles fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleListResponse"
              examples:
                bundles:
                  $ref: "#/components/examples/BundleListResponseExample"
    post:
      tags: [Bundles]
      summary: Create bundle package
      description: Adds a wholesale bundle package to the catalog. Admins may create hidden or disabled packages before release.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BundlePackageRequest"
            examples:
              createPackage:
                $ref: "#/components/examples/BundlePackageRequestExample"
      responses:
        "200":
          description: Bundle package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleResponse"
              examples:
                packageCreated:
                  $ref: "#/components/examples/BundleResponseExample"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /bundles/{bundleId}:
    parameters:
      - $ref: "#/components/parameters/BundleId"
    get:
      tags: [Bundles]
      summary: Get bundle package
      responses:
        "200":
          description: Bundle package fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleResponse"
              examples:
                package:
                  $ref: "#/components/examples/BundleResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [Bundles]
      summary: Update bundle package
      description: Updates package details, customer visibility, and active/paused/disabled state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BundlePackageUpdateRequest"
            examples:
              updatePackage:
                $ref: "#/components/examples/BundlePackageUpdateRequestExample"
      responses:
        "200":
          description: Bundle package updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleResponse"
              examples:
                packageUpdated:
                  $ref: "#/components/examples/BundleResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-mtn_bds_session
      description: HTTP-only, Secure, SameSite=Strict signed JWT session cookie. Local development may use mtn_bds_session over http.
  parameters:
    CustomerId:
      name: customerId
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: One-based page number used by paginated tables.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Number of rows per page. Infinite scrolling uses this as the fetch batch size.
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
      description: Case-insensitive search term applied to human-readable table fields.
...
    DateFrom:
      name: dateFrom
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Inclusive start date for date-filtered table rows. For /overview, used when revenuePeriod is custom.
    DateTo:
      name: dateTo
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Inclusive end date for date-filtered table rows. For /overview, used when revenuePeriod is custom.
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            validationError:
              $ref: "#/components/examples/ErrorResponseExample"
  examples:
...
    CustomerListResponseExample:
      summary: Paginated customer account table
      value:
        success: true
        message: Records fetched successfully
        data:
          - id: cus-wavenet
            businessName: WaveNet
            registrationNumber: UG-BR-48290
            businessEmail: business@wavenet.ug
            businessPhone: "+256772100200"
            contactPerson: Sarah Namuli
            email: operations@wavenet.ug
            phone: "+256772100201"
            apnName: wavenet.mtn.ug
            apnId: APN-1092
            primaryMsisdns: ["+256772990001", "+256772990002"]
            secondaryCount: 430
            bundlePurchases: 350
            totalSpendUgx: 18500000
            status: active
            createdAt: "2026-04-18T09:00:00+03:00"
        meta:
          page: 1
          limit: 10
          total: 1
          totalPages: 1
          hasNextPage: false
          hasPreviousPage: false
    CustomerRegistrationRequestExample:
      summary: Register customer account
      value:
        businessName: WaveNet Uganda
        registrationNumber: UG-BR-1029
        businessEmail: accounts@wavenet.ug
        businessPhone: "+256772991000"
        contactPerson: Sarah Namuli
        contactEmail: sarah.namuli@wavenet.ug
        contactPhone: "+256772991001"
        apnName: wavenet.mtn.ug
        apnId: APN-1092
        primaryMsisdn: "+256772990001"
    CustomerRegistrationResponseExample:
      summary: Customer registered
      value:
        success: true
        message: Customer registered successfully
        data:
          customer:
            id: cus-wavenet
            businessName: WaveNet Uganda
            registrationNumber: UG-BR-1029
            businessEmail: accounts@wavenet.ug
            businessPhone: "+256772991000"
            contactPerson: Sarah Namuli
            email: sarah.namuli@wavenet.ug
            phone: "+256772991001"
            apnName: wavenet.mtn.ug
            apnId: APN-1092
            primaryMsisdns: ["+256772990001"]
            secondaryCount: 0
            bundlePurchases: 0
            totalSpendUgx: 0
            status: pending
            createdAt: "2026-04-22T09:00:00+03:00"
          validation:
            msisdn: "+256772990001"
            accepted: true
            reason: MSISDN is valid for this APN.
            apnIds: [APN-1092]
            registeredApnId: APN-1092
            provisioningAction: addSubscriber
          activation:
            activationToken: act_demo
            activationUrl: /auth/activate?token=act_demo
            expiresAt: "2026-04-23T09:00:00+03:00"
            deliveryChannels: [business_email, contact_email]
    CustomerResponseExample:
      summary: Customer details
      value:
        success: true
        message: Operation completed successfully
        data:
          id: cus-wavenet
          businessName: WaveNet
          registrationNumber: UG-BR-48290
          businessEmail: business@wavenet.ug
          businessPhone: "+256772100200"
          contactPerson: Sarah Namuli
          email: operations@wavenet.ug
          phone: "+256772100201"
          apnName: wavenet.mtn.ug
          apnId: APN-1092
          primaryMsisdns: ["+256772990001", "+256772990002"]
          secondaryCount: 430
          bundlePurchases: 350
          totalSpendUgx: 18500000
          status: active
          createdAt: "2026-04-18T09:00:00+03:00"
    CustomerUpdateRequestExample:
      summary: Update customer contact details
      value:
        businessEmail: billing@wavenet.ug
        businessPhone: "+256772991009"
        contactPerson: Sarah Namuli
        contactEmail: sarah.namuli@wavenet.ug
        contactPhone: "+256772991001"
    CustomerStatusChangeRequestExample:
      summary: Deactivate customer account
      value:
        status: deactivated
        reason: Customer requested suspension.
        supportingNote: Approved by account manager.
    CustomerDeactivatedResponseExample:
      summary: Customer deactivated
      value:
        success: true
        message: Customer status updated successfully
        data:
          id: cus-wavenet
          businessName: WaveNet
          registrationNumber: UG-BR-48290
          businessEmail: business@wavenet.ug
          businessPhone: "+256772100200"
          contactPerson: Sarah Namuli
          email: operations@wavenet.ug
          phone: "+256772100201"
          apnName: wavenet.mtn.ug
          apnId: APN-1092
          primaryMsisdns: ["+256772990001", "+256772990002"]
          secondaryCount: 430
          bundlePurchases: 350
          totalSpendUgx: 18500000
          status: deactivated
          deactivationReason: Customer requested suspension.
          createdAt: "2026-04-18T09:00:00+03:00"
    PrimaryMsisdnRequestExample:
      summary: Add primary number
      value:
        primaryMsisdn: "+256772990003"
    PrimaryMsisdnResponseExample:
      summary: Primary number added
      value:
        success: true
        message: Primary MSISDN added successfully
        data:
          customer:
            id: cus-wavenet
            businessName: WaveNet
            registrationNumber: UG-BR-48290
            businessEmail: business@wavenet.ug
            businessPhone: "+256772100200"
            contactPerson: Sarah Namuli
            email: operations@wavenet.ug
            phone: "+256772100201"
            apnName: wavenet.mtn.ug
            apnId: APN-1092
            primaryMsisdns: ["+256772990001", "+256772990002", "+256772990003"]
            secondaryCount: 430
            bundlePurchases: 350
            totalSpendUgx: 18500000
            status: active
            createdAt: "2026-04-18T09:00:00+03:00"
          validation:
            msisdn: "+256772990003"
            accepted: true
            reason: MSISDN is valid for this APN.
            apnIds: [APN-1092]
            registeredApnId: APN-1092
            provisioningAction: addSubscriber
...
  schemas:
...
    CustomerRegistrationRequest:
      type: object
      required:
        - businessName
        - registrationNumber
        - businessEmail
        - businessPhone
        - contactPerson
        - contactEmail
        - contactPhone
        - apnName
        - apnId
        - primaryMsisdn
      properties:
        businessName:
          type: string
        registrationNumber:
          type: string
        businessEmail:
          type: string
          format: email
        businessPhone:
          $ref: "#/components/schemas/UgandaPhoneNumber"
        contactPerson:
          type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          $ref: "#/components/schemas/UgandaPhoneNumber"
        apnName:
          type: string
        apnId:
          type: string
        primaryMsisdn:
          $ref: "#/components/schemas/UgandaPhoneNumber"
...
    CustomerUpdateRequest:
      type: object
      properties:
        businessName:
          type: string
        businessEmail:
          type: string
          format: email
        businessPhone:
          $ref: "#/components/schemas/UgandaPhoneNumber"
        contactPerson:
          type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          $ref: "#/components/schemas/UgandaPhoneNumber"
        apnName:
          type: string
        apnId:
          type: string
    CustomerStatusChangeRequest:
      type: object
      required: [status, reason]
      properties:
        status:
          type: string
          enum: [active, deactivated]
        reason:
          type: string
        supportingNote:
          type: string
...
    PrimaryMsisdnRequest:
      type: object
      required: [primaryMsisdn]
      properties:
        primaryMsisdn:
          $ref: "#/components/schemas/UgandaPhoneNumber"
...
    CustomerResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Customer"
...
    CustomerListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Customer"
    CustomerRegistrationResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [customer, validation, activation]
              properties:
                customer:
                  $ref: "#/components/schemas/Customer"
                validation:
                  $ref: "#/components/schemas/MsisdnValidationResult"
                activation:
                  $ref: "#/components/schemas/CustomerActivationNotice"
...
    PrimaryMsisdnResponse:
      $ref: "#/components/schemas/CustomerRegistrationResponse"
...